# Conflicts:
#	src/main/java/com/shows/as/domain/classes/Showscom.java
#	src/main/java/com/shows/as/domain/classes/Test.java
#	src/main/resources/hibernate.cfg.xml

+++ src/main/java/com/shows/as/domain/adapters/BankServiceAdapter.java
@@ -5,12 +5,9 @@ import com.shows.as.service.BankService;
 
 import java.util.Date;
 
-/**
- * Created by victor on 12/01/2016.
- */
 public class BankServiceAdapter implements IBankServiceAdapter {
 
-    public Boolean autoritza(String dni, Integer codiB, String numCompte, Float preu, Integer codiBancShows, String numCompteShows, Date dAvui){
+    public Boolean autoritza(String dni, Integer codiB, String numCompte, Double preu, Integer codiBancShows, String numCompteShows, Date dAvui){
        ServiceLocator  s = ServiceLocator.getInstance();
         BankService b = s.getBankService();
         return b.autoritza(dni,codiB,numCompte,preu,codiBancShows,numCompteShows,dAvui);
